move client_configuration_finder to client_configuration
